Use cdpetron always from the project's directory.
This was premature optimization from the cdpetron's early days, in
reality we ALWAYS use it from inside the branch, so it doesn't make
sense to specify it (even it was annoying to have that "." always passed
to cdpetron.

So I removed that, which led to a bunch of nice simplifications.